+++ src/Block/Renderer/BlockQuoteRenderer.php
@@ -35,10 +35,7 @@ class BlockQuoteRenderer implements BlockRendererInterface
             throw new \InvalidArgumentException('Incompatible block type: ' . \get_class($block));
         }
 
-        $attrs = [];
-        foreach ($block->getData('attributes', []) as $key => $value) {
-            $attrs[$key] = Xml::escape($value);
-        }
+        $attrs = $block->getData('attributes', []);
 
         $filling = $htmlRenderer->renderBlocks($block->children());
         if ($filling === '') {