+++ src/Inline/Renderer/ImageRenderer.php
@@ -42,16 +42,13 @@ class ImageRenderer implements InlineRendererInterface, ConfigurationAwareInterf
             throw new \InvalidArgumentException('Incompatible inline type: ' . \get_class($inline));
         }
 
-        $attrs = [];
-        foreach ($inline->getData('attributes', []) as $key => $value) {
-            $attrs[$key] = Xml::escape($value);
-        }
+        $attrs = $inline->getData('attributes', []);
 
         $forbidUnsafeLinks = $this->config->getConfig('safe') || !$this->config->getConfig('allow_unsafe_links');
         if ($forbidUnsafeLinks && RegexHelper::isLinkPotentiallyUnsafe($inline->getUrl())) {
             $attrs['src'] = '';
         } else {
-            $attrs['src'] = Xml::escape($inline->getUrl());
+            $attrs['src'] = $inline->getUrl();
         }
 
         $alt = $htmlRenderer->renderInlines($inline->children());
@@ -59,7 +56,7 @@ class ImageRenderer implements InlineRendererInterface, ConfigurationAwareInterf
         $attrs['alt'] = \preg_replace('/\<[^>]*\>/', '', $alt);
 
         if (isset($inline->data['title'])) {
-            $attrs['title'] = Xml::escape($inline->data['title']);
+            $attrs['title'] = $inline->data['title'];
         }
 
         return new HtmlElement('img', $attrs, '', true);